add server side logic for starting a meeting with no image

+++ backend/server.js
@@ -273,38 +273,37 @@ io.on('connection', (client) => {
       })
       .then((id) => {
         fs.mkdir(`meeting_files/${id}`, () => { //makes a new directory for the meeting
-          //Check if pdf
-          if (data.file.name.search(/\.pdf$/ig) !== -1) {
+          if (data.filename) { //Only save the file if one exists, otherwise just have an empty folder
+            //Check if pdf
+            if (data.file.name.search(/\.pdf$/ig) !== -1) {
 
-            console.log(data.file.name);
-            fs.writeFile(`meeting_files/${id}/${data.file.name}`, data.file.payload, (err) => {
-              if (err) {
-                console.log('problem');
-                throw err;
-              }
-              console.log('The file has been saved!');
+              console.log(data.file.name);
+              fs.writeFile(`meeting_files/${id}/${data.file.name}`, data.file.payload, (err) => {
+                if (err) {
+                  console.log('problem');
+                  throw err;
+                }
+                console.log('The file has been saved!');
 
-              let pdfImage = new PDFImage(`meeting_files/${id}/${data.file.name}`);
+                let pdfImage = new PDFImage(`meeting_files/${id}/${data.file.name}`);
 
-              // console.log(pdfImage);
+                // console.log(pdfImage);
 
-              pdfImage.convertPage(0).then(function(imagePath) {
-                // 0-th page (first page) of the slide.pdf is available as slide-0.png
-                fs.existsSync("/tmp/slide-0.png") // => true\
-                console.log(imagePath);
-              })
-                .catch((error) => {
-                  console.log(error);
+                pdfImage.convertPage(0).then(function(imagePath) {
+                  // 0-th page (first page) of the slide.pdf is available as slide-0.png
+                  fs.existsSync("/tmp/slide-0.png") // => true\
+                  console.log(imagePath);
                 })
-            }); //Note promisy this I if we want to wait for the upload to finish before creating meeting
-
-
-
-          } else {
-            fs.writeFile(`meeting_files/${id}/${data.file.name}`, data.file.payload, (err) => {
-              if (err) throw err;
-              console.log('The file has been saved!');
-            }); //Note promisy this I if we want to wait for the upload to finish before creating meeting
+                  .catch((error) => {
+                    console.log(error);
+                  })
+              }); //Note promisy this I if we want to wait for the upload to finish before creating meeting
+            } else {
+              fs.writeFile(`meeting_files/${id}/${data.file.name}`, data.file.payload, (err) => {
+                if (err) throw err;
+                console.log('The file has been saved!');
+              }); //Note promisy this I if we want to wait for the upload to finish before creating meeting
+            }
           }
         });
         const promiseArray = [];
@@ -374,30 +373,43 @@ io.on('connection', (client) => {
     }
     console.log("Looking for", `meeting_files/${data.meetingId}/${meetingDetails.link_to_initial_doc}`);
 
-    let img;
-    if (meetingDetails.link_to_initial_doc.search(/\.pdf$/ig) !== -1) {
-      img = meetingDetails.link_to_initial_doc.split(/\.pdf$/ig)[0] + "-0.png";
-    } else {
-      img = meetingDetails.link_to_initial_doc;
-    }
-
-    fs.readFile(`meeting_files/${data.meetingId}/${img}`, (err, image) => {
-      if (err) {
-        console.error;
-        image = "";
+    let img = "";
+    if (meetingDetails.link_to_initial_doc) {
+      if (meetingDetails.link_to_initial_doc.search(/\.pdf$/ig) !== -1) {
+        img = meetingDetails.link_to_initial_doc.split(/\.pdf$/ig)[0] + "-0.png";
+      } else {
+        img = meetingDetails.link_to_initial_doc;
       }
-      console.log("sending these pixels");
-      console.log(meetingDetails.userPixels);
+      fs.readFile(`meeting_files/${data.meetingId}/${img}`, (err, image) => {
+        if (err) {
+          console.error;
+          image = "";
+        }
+        console.log("sending these pixels");
+        console.log(meetingDetails.userPixels);
 
+        db.fetchUsersMeetingsByIds(data.user.id, data.meetingId)
+          .then((res) => {
+
+            client.emit('enteredMeeting', { meeting: meetingDetails, notes: res[0].notes, pixels: meetingDetails.userPixels, image: "data:image/jpg;base64," + image.toString("base64") });
+
+            client.join(data.meetingId);
+            io.to(data.meetingId).emit('newParticipant', (data.user));
+          });
+      });
+    } else {
       db.fetchUsersMeetingsByIds(data.user.id, data.meetingId)
         .then((res) => {
 
-          client.emit('enteredMeeting', { meeting: meetingDetails, notes: res[0].notes, pixels: meetingDetails.userPixels, image: "data:image/jpg;base64," + image.toString("base64") });
+          client.emit('enteredMeeting', { meeting: meetingDetails, notes: res[0].notes, pixels: meetingDetails.userPixels, image: "" });
 
           client.join(data.meetingId);
           io.to(data.meetingId).emit('newParticipant', (data.user));
-        })
-    });
+        });
+    }
+
+
+
   });
 
   client.on('saveDebouncedNotes', (data) => {
